Factorise le rendu des champs d'une réponse vers un composant dédié
