+++ web/src/pages/ResponsePage.vue
@@ -15,7 +15,6 @@ import { useApiFetch } from "../utils/data-fetching.ts"
 import type { SurveyField, ImageItem } from "@shared-types/survey"
 import SurveyRenderer from "@shared-components/SurveyRenderer.vue"
 import ImageViewer from "@shared-components/ImageViewer.vue"
-import { resolveFieldValue } from "@shared-utils/survey"
 import { storeToRefs } from "pinia"
 import { useRootStore } from "../stores/root.ts"
 import MapField from "../components/MapField.vue"
@@ -24,6 +23,7 @@ import ConfirmDeleteModal from "../components/SurveyBuilder/ConfirmDeleteModal.v
 import { useToastStore } from "../stores/toast"
 import { useRouter } from "vue-router"
 import FollowUpsTable from "../components/ResponsePage/FollowUpsTable.vue"
+import ResponseFieldsSection from "../components/ResponsePage/ResponseFieldsSection.vue"
 
 const router = useRouter()
 const route = useRoute()
@@ -72,43 +72,6 @@ watch([response, isFetching], async ([val, fetching]) => {
 
 const confirmDeleteOpened = ref(false)
 const toast = useToastStore()
-
-const fieldLabel = (fieldId: string): string =>
-  response.value?.survey.jsonSchema.fields.find(
-    (f: SurveyField) => f.id === fieldId
-  )?.label ?? fieldId
-
-const resolveValue = (fieldId: string, raw: unknown): string => {
-  const field = response.value?.survey.jsonSchema.fields.find(
-    (f: SurveyField) => f.id === fieldId
-  )
-  return resolveFieldValue(field, raw, surveyVocabularies.value)
-}
-
-const resolveSubFieldValue = (subField: SurveyField, raw: unknown): string =>
-  resolveFieldValue(subField, raw, surveyVocabularies.value)
-
-const isArrayField = (fieldId: string): boolean =>
-  response.value?.survey.jsonSchema.fields.find(
-    (f: SurveyField) => f.id === fieldId
-  )?.ui?.widget === "array"
-
-const isImageField = (fieldId: string): boolean =>
-  response.value?.survey.jsonSchema.fields.find(
-    (f: SurveyField) => f.id === fieldId
-  )?.ui?.widget === "image"
-
-const getSubFields = (fieldId: string): SurveyField[] =>
-  response.value?.survey.jsonSchema.fields.find(
-    (f: SurveyField) => f.id === fieldId
-  )?.fields ?? []
-
-const imageSrc = (item: ImageItem): string | null => {
-  if ("type" in item) return null
-  if ("file" in item) return `data:image/jpeg;base64,${item.file}`
-  if (item.thumbnail) return `data:image/jpeg;base64,${item.thumbnail}`
-  return null
-}
 
 const viewerOpen = ref(false)
 const viewerImages = ref<ImageItem[]>([])
@@ -172,103 +135,11 @@ const onConfirmDelete = async () => {
 
       <div class="grid grid-cols-12 gap-4">
         <div class="col-span-12 sm:col-span-6 md:col-span-7 lg:col-span-8">
-          <div v-for="entry in Object.entries(response.data)" :key="entry[0]">
-            <p class="fr-text--sm font-bold text-stone-500 mb-0!">
-              {{ fieldLabel(entry[0]) }}
-            </p>
-
-            <!-- Array field -->
-            <template v-if="isArrayField(entry[0]) && Array.isArray(entry[1])">
-              <p v-if="!entry[1].length" class="italic text-stone-500 mb-0!">
-                Non renseigné
-              </p>
-              <p v-else class="font-medium mb-2! text-stone-500">
-                {{ entry[1].length }} entrée(s) :
-              </p>
-              <div class="grid grid-cols-12 gap-4">
-                <div
-                  v-for="(item, idx) in (entry[1] as Record<string, unknown>[])"
-                  :key="idx"
-                  class="border border-slate-200 rounded p-3 bg-slate-50 col-span-12 md:col-span-6 lg:col-span-4"
-                >
-                  <div
-                    v-for="subField in getSubFields(entry[0])"
-                    :key="subField.id"
-                  >
-                    <p class="fr-text--sm text-stone-400 mb-0!">
-                      {{ subField.label }}
-                    </p>
-                    <p
-                      class="font-medium mb-0!"
-                      v-if="resolveSubFieldValue(subField, item[subField.id])"
-                    >
-                      {{ resolveSubFieldValue(subField, item[subField.id]) }}
-                    </p>
-                    <p class="italic text-stone-500 mb-0!" v-else>
-                      Non renseigné
-                    </p>
-                  </div>
-                </div>
-              </div>
-            </template>
-
-            <!-- Champ images -->
-            <template
-              v-else-if="isImageField(entry[0]) && Array.isArray(entry[1])"
-            >
-              <p v-if="!entry[1].length" class="italic text-stone-500 mb-0!">
-                Non renseigné
-              </p>
-              <div v-else class="grid grid-cols-7 gap-2 my-2">
-                <div
-                  v-for="(img, idx) in (entry[1] as ImageItem[])"
-                  :key="idx"
-                  class="group aspect-square rounded overflow-hidden border border-slate-200 cursor-pointer relative"
-                  @click="openViewer(entry[1] as ImageItem[], idx)"
-                >
-                  <img
-                    v-if="imageSrc(img)"
-                    :src="imageSrc(img)!"
-                    class="w-full h-full object-cover"
-                    alt=""
-                  />
-                  <div
-                    v-else
-                    class="w-full h-full bg-slate-100 flex items-center justify-center"
-                  >
-                    <v-icon
-                      name="ri-image-line"
-                      scale="2"
-                      class="text-slate-400"
-                    />
-                  </div>
-                  <div
-                    class="absolute inset-0 bg-black/0 group-hover:bg-black/30 transition-colors flex items-center justify-center"
-                  >
-                    <v-icon
-                      name="ri-search-eye-line"
-                      color="white"
-                      scale="1.5"
-                      class="text-white opacity-0 group-hover:opacity-100 transition-opacity"
-                    />
-                  </div>
-                </div>
-              </div>
-            </template>
-
-            <!-- All other fields -->
-            <template v-else>
-              <p
-                class="font-medium mb-0!"
-                v-if="resolveValue(entry[0], entry[1])"
-              >
-                {{ resolveValue(entry[0], entry[1]) }}
-              </p>
-              <p class="italic text-stone-500 mb-0!" v-else>Non renseigné</p>
-            </template>
-
-            <hr class="mt-2! mb-0!" />
-          </div>
+          <ResponseFieldsSection
+            :response="response"
+            :surveyCodes="surveyCodes"
+            @open-viewer="openViewer"
+          />
         </div>
         <!-- Preview -->
         <div class="col-span-12 sm:col-span-6 md:col-span-5 lg:col-span-4 mb-4">
